# docs/openapi.yaml
openapi: 3.0.3
info:
  title: VulScan API
  description: 企业级Web漏洞扫描平台API文档
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development server

tags:
  - name: Auth
    description: 用户认证相关接口
  - name: Targets
    description: 目标管理相关接口
  - name: ScanTasks
    description: 扫描任务相关接口
  - name: Vulnerabilities
    description: 漏洞管理相关接口
  - name: Reports
    description: 报告管理相关接口
  - name: Metrics
    description: 监控指标相关接口

components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: datetime
      required:
        - id
        - username
        - email
        - role
        - is_active

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required:
        - email
        - password

    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        user:
          $ref: '#/components/schemas/User'

    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required:
        - refresh_token

    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required:
        - username
        - email
        - password

    Target:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        url:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
        created_at:
          type: string
          format: datetime
        updated_at:
          type: string
          format: datetime
      required:
        - id
        - name
        - url
        - status

    CreateTargetRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        description:
          type: string
      required:
        - name
        - url

    ScanTask:
      type: object
      properties:
        id:
          type: string
          format: uuid
        target_id:
          type: string
          format: uuid
        target_name:
          type: string
        status:
          type: string
        scan_type:
          type: string
        priority:
          type: string
        progress:
          type: integer
        started_at:
          type: string
          format: datetime
          nullable: true
        completed_at:
          type: string
          format: datetime
          nullable: true
        created_at:
          type: string
          format: datetime
      required:
        - id
        - target_id
        - status
        - scan_type
        - priority
        - progress

    CreateScanTaskRequest:
      type: object
      properties:
        target_id:
          type: string
          format: uuid
        scan_type:
          type: string
          default: full
        priority:
          type: string
          default: medium
        plugins:
          type: array
          items:
            type: string
      required:
        - target_id

    Vulnerability:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        target_id:
          type: string
          format: uuid
        plugin_name:
          type: string
        severity:
          type: string
        title:
          type: string
        description:
          type: string
        payload:
          type: string
          nullable: true
        proof:
          type: string
          nullable: true
        remediation:
          type: string
        cve:
          type: string
          nullable: true
        cvss_score:
          type: number
          format: float
          nullable: true
        created_at:
          type: string
          format: datetime
      required:
        - id
        - task_id
        - target_id
        - plugin_name
        - severity
        - title
        - description
        - remediation

    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        task_name:
          type: string
        status:
          type: string
        summary:
          type: object
        generated_at:
          type: string
          format: datetime
          nullable: true
        created_at:
          type: string
          format: datetime
      required:
        - id
        - task_id
        - status
        - summary

    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items:
            type: object
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
        timestamp:
          type: string
          format: datetime

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []

paths:
  /auth/login:
    post:
      tags:
        - Auth
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 认证失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/register:
    post:
      tags:
        - Auth
      summary: 用户注册
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 请求错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/refresh:
    post:
      tags:
        - Auth
      summary: 刷新令牌
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 刷新令牌无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/logout:
    post:
      tags:
        - Auth
      summary: 用户登出
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 登出成功

  /targets:
    get:
      tags:
        - Targets
      summary: 获取目标列表
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
        - name: page_size
          in: query
          type: integer
          default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Target'
    post:
      tags:
        - Targets
      summary: 创建目标
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTargetRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'

  /targets/{id}:
    get:
      tags:
        - Targets
      summary: 获取目标详情
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
        '404':
          description: 目标不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Targets
      summary: 更新目标
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTargetRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Target'
    delete:
      tags:
        - Targets
      summary: 删除目标
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: 删除成功
        '404':
          description: 目标不存在

  /scan-tasks:
    get:
      tags:
        - ScanTasks
      summary: 获取扫描任务列表
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
        - name: page_size
          in: query
          type: integer
          default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ScanTask'
    post:
      tags:
        - ScanTasks
      summary: 创建扫描任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScanTaskRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanTask'

  /scan-tasks/{id}:
    get:
      tags:
        - ScanTasks
      summary: 获取扫描任务详情
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanTask'
    put:
      tags:
        - ScanTasks
      summary: 更新扫描任务
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanTask'
    delete:
      tags:
        - ScanTasks
      summary: 删除扫描任务
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: 删除成功

  /scan-tasks/{id}/cancel:
    put:
      tags:
        - ScanTasks
      summary: 取消扫描任务
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: 取消成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScanTask'

  /vulnerabilities:
    get:
      tags:
        - Vulnerabilities
      summary: 获取漏洞列表
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
        - name: page_size
          in: query
          type: integer
          default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Vulnerability'

  /vulnerabilities/{id}:
    get:
      tags:
        - Vulnerabilities
      summary: 获取漏洞详情
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vulnerability'
    delete:
      tags:
        - Vulnerabilities
      summary: 删除漏洞
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: 删除成功

  /reports:
    get:
      tags:
        - Reports
      summary: 获取报告列表
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
        - name: page_size
          in: query
          type: integer
          default: 10
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Report'

  /reports/{id}:
    get:
      tags:
        - Reports
      summary: 获取报告详情
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Report'
    delete:
      tags:
        - Reports
      summary: 删除报告
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: 删除成功

  /metrics:
    get:
      tags:
        - Metrics
      summary: 获取监控指标
      security: []
      responses:
        '200':
          description: 获取成功
          content:
            text/plain:
              schema:
                type: string

  /health:
    get:
      tags:
        - Metrics
      summary: 健康检查
      security: []
      responses:
        '200':
          description: 服务正常
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: datetime